+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 Формат основан на [стандарте формата CHANGELOG](https://keepachangelog.com/en/1.0.0/),
 и придерживается [правил версионирования](https://semver.org/spec/v2.0.0.html).
 
+- Изменено:
+  - Переименован класс QueryBuilder на MySql\Connection
+
 ## [ [0.4.2](https://github.com/C-o-d-eCraft/framework/releases/tag/0.4.2) ] - 20.10.2024
 - Реализовано:
   - Реализована возможность добавлять мидлвеары на отдельные маршруты ресурсов